+++ src/tasks/alacritty.rs
@@ -2,12 +2,14 @@ use lib::task::{self, Status, Task};
 use utils;
 
 pub fn task() -> Task {
-    Task { sync, update }
+    Task {
+        name: "alacritty".to_string(),
+        sync,
+        update,
+    }
 }
 
 fn sync() -> task::Result {
-    println!("alacritty: syncing ...");
-
     let src = utils::env::home_dir().join(".dotfiles/config/alacritty.yml");
     let dest = utils::env::home_dir().join(".config/alacritty/alacritty.yml");