+++ src/tasks/shfmt.rs
@@ -1,18 +1,17 @@
 use regex::Regex;
 
-use lib::ghrtask::GHRTask;
+use lib::{
+    ghrtask::GHRTask,
+    task::{self, Task},
+};
 use utils::github::Asset;
 use utils::golang::{arch, os};
 
-pub fn sync() {
-    match GHR_TASK.sync() {
-        _ => {}
-    }
-}
-
-pub fn update() {
-    match GHR_TASK.update() {
-        _ => {}
+pub fn task() -> Task {
+    Task {
+        name: "shfmt".to_string(),
+        sync,
+        update,
     }
 }
 
@@ -33,7 +32,15 @@ fn asset_filter(asset: &Asset) -> bool {
     re.is_match(&asset.name)
 }
 
+fn sync() -> task::Result {
+    GHR_TASK.sync()
+}
+
 #[cfg_attr(feature = "cargo-clippy", allow(needless_pass_by_value))]
 fn trim_version(stdout: String) -> String {
     String::from(stdout.trim())
+}
+
+fn update() -> task::Result {
+    GHR_TASK.update()
 }